+++ resources/views/bootstrap-5/components/buttons/link-button.blade.php
@@ -1,6 +1,8 @@
 <a
     @if ($disabled === false)
-        href="{{ $url }}"
+        @if ($url !== null)
+            href="{{ $url }}"
+        @endif
         {{ $attributes->merge(['class' => 'btn btn-'.$variant.($size !== null ? ' btn-'.$size : ''), 'role' => 'button']) }}
     @else
         {{ $attributes->merge(['class' => 'disabled btn btn-'.$variant.($size !== null ? ' btn-'.$size : ''), 'role' => 'button', 'aria-disabled' => 'true', 'tabindex' => '-1']) }}