+++ src/Components/Buttons/Actions/Create.php
@@ -13,11 +13,10 @@ class Create extends LinkButton
     {
         $this->variant ??= 'primary';
 
-        $this->confirmVariant ??= 'variant';
-
         $this->text ??= Str::ucfirst(trans('action.add'));
 
         if ($this->confirm !== null) {
+            $this->confirmVariant ??= 'primary';
             $this->confirmId = 'create-'.($this->confirmId ?? Str::random(32));
         }
     }